山手線風テーマ仮実装
